Corrige grafico de KM vazio: baseline na 1a leitura automatica e KM sem arredondar
O grafico de KM por mes soma deltas entre leituras consecutivas de GPS, mas a
primeira leitura confirmada de um carro nunca tinha uma leitura anterior pra
comparar - por isso o dashboard ficou vazio na estreia do relatorio automatico.
Agora salvar_leitura_automatica grava uma leitura baseline (KM 0, um dia antes)
quando o carro ainda nao tem nenhuma leitura confirmada, e um comando
`flask backfill-gps-baseline` cobre retroativamente os carros que ja ficaram
presos nesse estado antes da correcao.

Tambem para de arredondar o KM do relatorio automatico (round() fazia deltas
pequenos, como os 0,03 km de um dia parado, sumirem do carro pra sempre).
current_km/extracted_km/confirmed_km viram FLOAT (migration incluida), com um
filtro Jinja `km` pra formatar no padrao brasileiro nas telas que exibem esses
valores.

+++ models.py
@@ -41,7 +41,7 @@ class Car(db.Model):
     plate = db.Column(db.String(20), unique=True, nullable=False)
     model = db.Column(db.String(120))
     year = db.Column(db.Integer)
-    current_km = db.Column(db.Integer, default=0, nullable=False)
+    current_km = db.Column(db.Float, default=0, nullable=False)
     active = db.Column(db.Boolean, default=True, nullable=False)
     created_at = db.Column(db.DateTime, default=datetime.utcnow)
 
@@ -281,7 +281,7 @@ class GPSReading(db.Model):
     car_id = db.Column(db.Integer, db.ForeignKey("cars.id"), nullable=False)
     image_filename = db.Column(db.String(255), nullable=False)
 
-    extracted_km = db.Column(db.Integer)
+    extracted_km = db.Column(db.Float)
     extracted_max_speed = db.Column(db.Integer)
     extracted_moving_time_minutes = db.Column(db.Integer)
     raw_ai_response = db.Column(db.Text)
@@ -290,7 +290,7 @@ class GPSReading(db.Model):
     status = db.Column(db.String(15), nullable=False, default="pending_review")
     # "pending_review" | "confirmed" | "rejected"
 
-    confirmed_km = db.Column(db.Integer)
+    confirmed_km = db.Column(db.Float)
     confirmed_max_speed = db.Column(db.Integer)
     confirmed_moving_time_minutes = db.Column(db.Integer)
     reading_date = db.Column(db.Date)
